test: enable or remove disabled tests
Some tests will be added in next commit
Remove all product related redux code
Cause Product state moved to Formik
