+++ src/components/pages/add-order-page/AddOrderPage.test.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import AddOrderPage from './AddOrderPage'
 
-describe.skip('AddOrderPage', () => {
+describe('AddOrderPage', () => {
   const container = document.createElement('div')
   document.body.appendChild(container)
   const defaultProps = {
@@ -19,36 +19,9 @@ describe.skip('AddOrderPage', () => {
     ReactDOM.render(<AddOrderPage {...localProps} />, container)
   }
 
-  function querySelectorAll(...params) {
-    return container.querySelectorAll(...params)
-  }
-
   afterEach(() => ReactDOM.unmountComponentAtNode(container))
 
   it('renders without crashing', () => {
     render()
   })
-
-  it('renders 0 products', () => {
-    render({
-      products: [],
-    })
-    expect(querySelectorAll('.products > *')).toHaveLength(0)
-  })
-
-  it(
-    'calls addProduct callback on click on add product button ' +
-      'without arguments',
-    () => {
-      const onAddProduct = jest.fn()
-      render({
-        onAddProduct,
-      })
-
-      const addProductButton = querySelectorAll('.btn-add-product')[0]
-      addProductButton.click()
-
-      expect(onAddProduct).toHaveBeenCalledWith()
-    }
-  )
 })
