+++ src/components/product/Product.stories.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 import { action } from '@storybook/addon-actions'
-import { Product } from './Product'
+import Product from './Product'
 import { OPTIONS_MODES } from '../../models/product_model'
 import { Formik } from 'formik'
