+++ src/components/product/Product.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import { Product } from './Product'
+import Product from './Product'
 import { OPTIONS_MODES } from '../../models/product_model'
 import { Formik } from 'formik'
