--- src/components/product/index.js
@@ -1,3 +1,0 @@
-import ConnectedProduct from './Product'
-
-export default ConnectedProduct
